ai_engine/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS and middleware
python-multipart==0.0.6

# Future AI/ML dependencies (for upcoming modules)
# langchain==0.1.0
# chromadb==0.4.22
# sentence-transformers==2.2.2
# neo4j==5.15.0
# openai==1.7.2
# huggingface-hub==0.20.2

# Utilities
requests==2.31.0
chromadb
sentence-transformers
